+++ src/Components/Filters/styles.module.scss
@@ -1,0 +1,24 @@
+@import '../../assets/styles/palette.scss';
+
+.filtersWrapper {
+  background-color: $yellow;
+  border: 7px solid $yellow;
+
+  & > .filters {
+    background-color: $white;
+    border: 5px solid $light-blue;
+    & > select {
+      font-size: 0.8rem;
+    }
+  }
+}
+
+.pokeRadio {
+  position: absolute;
+  opacity: 0;
+  width: 0;
+  height: 0;
+  & + img {
+    cursor: pointer;
+  }
+}
