--- src/Components/Loader.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Loader = () => {
-  return (
-    <div className='loader rounded-circle my-3' />
-  )
-}
