+++ src/Components/Loader/Loader.tsx
@@ -1,0 +1,7 @@
+import style from './styles.module.scss'
+
+export const Loader = () => {
+  return (
+    <div className={`${style.loader} rounded-circle my-3`} />
+  )
+}
