+++ src/assets/styles/palette.scss
@@ -1,0 +1,6 @@
+$light-orange: #FDEAB8;
+$light-blue: #85B6EF;
+$white: #F0F6F6;
+$yellow: #F7DC65;
+$orange: #F49742;
+$red: #DA4241;
