+++ src/declarations.d.ts
@@ -1,2 +1,3 @@
 declare module '*.jpg'
 declare module '*.png'
+declare module '*.scss';
